tried my hardest to hit the stretch goal, but there is a small issue in my logic

+++ src/piggiebank/Dime.java
@@ -2,7 +2,7 @@ package piggiebank;
 
 public class Dime extends AbstractMoney{
     
-    public Dime(double amt) {
+    public Dime(int amt) {
         super(amt, 0.10, "Dime");
     }
     public Dime() {
