+++ src/piggiebank/Dollar.java
@@ -2,7 +2,7 @@ package piggiebank;
 
 public class Dollar extends AbstractMoney{
     
-    public Dollar(double amt) {
+    public Dollar(int amt) {
         super(amt, 1, "Dollar");
     }
     public Dollar() {
