+++ src/piggiebank/Pennie.java
@@ -2,7 +2,7 @@ package piggiebank;
 
 public class Pennie extends AbstractMoney{
     
-    public Pennie(double amt) {
+    public Pennie(int amt) {
         super(amt, 0.01, "Pennie");
     }
     public Pennie() {
