+++ src/piggiebank/Quarter.java
@@ -2,7 +2,7 @@ package piggiebank;
 
 public class Quarter extends AbstractMoney{
     
-    public Quarter(double amt) {
+    public Quarter(int amt) {
         super(amt, 0.25, "Quarter");
     }
     public Quarter() {
